DispatchedTuples.jl

DispatchedTuples.jl defines one user-facing type: DispatchedTuple, and one user-facing method: dispatch. A DispatchedTuple is similar to a compile-time dictionary, that uses dispatch for the look-up.

DispatchedTuple takes a Tuple of Pairs, where the first field of the Pair (the "key") is an instance of the type you want to dispatch on. The second field of the Pair is the quantity (the "value", which can be anything) returned by dispatch.

Example

Here is an example in action

using DispatchedTuples
struct Foo end;
struct Bar end;
dtup = DispatchedTuple((
               Pair(Foo(), 1),
               Pair(Bar(), 2),
           ))

println(dispatch(dtup, Foo()))
println(dispatch(dtup, Bar()))

(1,)
(2,)

If a DispatchedTuple has duplicate keys, then all values are returned in the Tuple. Here's an example with duplicate keys:

using DispatchedTuples
struct Foo end
struct Bar end

dtup = DispatchedTuple((
               Pair(Foo(), 1),
               Pair(Foo(), 3),
               Pair(Bar(), 2),
           ))

println(dispatch(dtup, Foo()))
println(dispatch(dtup, Bar()))

(1, 3)
(2,)

The second (optional) argument to DispatchedTuple is a default value, which is returned for any unrecognized keys. If the default value is not given, and dispatch is called with a key it hasn't seen, an error is thrown.

For convenience, DispatchedTuple can alternatively take a Tuple of two-element Tuples.

API

DispatchedTuples.AbstractDispatchedTuple — Type

AbstractDispatchedTuple{T <: Tuple, D}

An abstract dispatch tuple type, for sub-typing dispatched tuples.

DispatchedTuples.DispatchedTuple — Type

DispatchedTuple(tup[, default_value])

A dispatch-able tuple.

DispatchedTuple takes a Tuple of Pairs, where the first field of the Pair (the "key") is an instance of the type you want to dispatch on. The second field of the Pair is the quantity (the "value", which can be anything) returned by dispatch.

If a DispatchedTuple has duplicate keys, then all values are returned in the Tuple.

The second (optional) argument to DispatchedTuple is a default value, which is returned for any unrecognized keys. If the default value is not given, and dispatch is called with a key it hasn't seen, an error is thrown.

For convenience, DispatchedTuple can alternatively take a Tuple of two-element Tuples.

DispatchedTuples.DispatchedTupleSet — Type

DispatchedTupleSet(tup[, default_value])

Similar to DispatchedTuple, except:

keys must be unique.

returns the value, and not a tuple of values.

throws an error in dispatch if keys are not unique.

DispatchedTuples.dispatch — Function

dispatch(::DispatchedTuple, type_instance)

Dispatch on the DispatchedTuple, based on the instance of the input type type_instance.

dispatch(::DispatchedTupleSet, type_instance)

Dispatch on the DispatchedTupleSet, based on the instance of the input type type_instance.