API

DispatchedTuples.AbstractDispatchedTuple — Type

AbstractDispatchedTuple{T <: Tuple, D}

An abstract dispatch tuple type, for sub-typing dispatched tuples.

DispatchedTuples.DispatchedTuple — Type

DispatchedTuple(tup[, default_value])

A dispatch-able tuple.

DispatchedTuple takes a Tuple of Pairs, where the first field of the Pair (the "key") is an instance of the type you want to dispatch on. The second field of the Pair is the quantity (the "value", which can be anything) returned by dispatch.

If a DispatchedTuple has non-unique keys, then all values are returned in the returned Tuple.

The second (optional) argument to DispatchedTuple is a default value, which is returned for any unrecognized keys. If the default value is not given, then dispatch on an unregistered key will return an empty tuple.

For convenience, DispatchedTuple can alternatively take a Tuple of two-element Tuples.

Example

using DispatchedTuples
struct Foo end;
struct Bar end;
struct Baz end;

dtup = DispatchedTuple((
   Pair(Foo(), 1),
   Pair(Foo(), 2),
   Pair(Bar(), 3),
));

dispatch(dtup, Foo()) # returns (1, 2)
dispatch(dtup, Bar()) # returns (3,)
dispatch(dtup, Baz()) # returns ()

DispatchedTuples.DispatchedSet — Type

DispatchedSet(tup[, default_value])

Similar to DispatchedTuple, except:

keys must be unique.

returns the value, and not a tuple of values.

throws an error in dispatch if the key is not unique (without a default value).

DispatchedTuples.dispatch — Function

dispatch(::DispatchedTuple, type_instance)

Dispatch on the DispatchedTuple, based on the instance of the input type type_instance.

dispatch(::DispatchedSet, type_instance)

Dispatch on the DispatchedSet, based on the instance of the input type type_instance.